+++ src/main/java/com/likeageek/randomizer/RandomEngine.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Random;
 
 public class RandomEngine {
-    public List<String> random(List<String> input, int seed) {
+    public List<String> random(List<String> input, long seed) {
         List<String> output = new ArrayList<>(input);
         Collections.shuffle(output, new Random(seed));
         return output;
