+++ src/main/java/com/likeageek/randomizer/shufflers/IShuffler.java
@@ -3,7 +3,7 @@ package com.likeageek.randomizer.shufflers;
 import java.util.Map;
 
 public interface IShuffler {
-    Map<String, String> shuffle(int seed);
+    Map<String, String> shuffle(long seed);
 
     void process(Map<String, String> shuffledEntries);
